refactor : test fixtures
